requirements.txt
asgiref==3.4.1
autopep8==1.6.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
cryptography==36.0.1
defusedxml==0.7.1
Django==4.0.1
django-allauth==0.44.0
django-bootstrap-modal-forms==2.2.0
django-crispy-forms==1.13.0
django-filter==21.1
django-formtools==2.3
django-widget-tweaks==1.4.12
idna==3.3
oauthlib==3.1.1
psycopg2==2.9.3
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.3.0
python3-openid==3.2.0
requests==2.27.1
requests-oauthlib==1.3.0
sqlparse==0.4.2
toml==0.10.2
urllib3==1.26.8
